Lots of small cumulative updates
Didnt even keep a mental list of these...
Really, just minor changes to the UI. I've been updating this library to
suit the needs of each project (so far, just one) it has been used on.
Which is to say, not very good coding behavior. there's certainly
nothing wrong with expanding a library/api in a project but Im really
just modifying it to suit my needs, which means it needs a complete
rewrite to be the most flexible (and not need changes!) In other words I
dont feel bothered to write commit messages for smoething that is going
to be replaced. mabye.