Added Scroll List, Prompt, and made more versatile
- Scroll list represents a region that displays a list of components
that can be scrolled through. It probably should extend UIPanel but it
doesn't and I am okay with that.
- Added CoordinatePair which pairs an object with some coordinates.
Handy forthe UIScroll list which needs to know where a component is. The
coordinates from the coord pair in the Scroll List represent where the
objects are rendered, meaning their normal static position Vector2i
thing is ignored completely. Really the best way to think about
ScrollLists are as a display screen, and the coord pair tells where on
the screen it is- whether it is even on screen or not. Similar to my
actual Screen class unrelated to all of this.
- All components are required to report a size now, even if they don't
really use size. This is to make them more uniform, and less breaky.
- UILabels now have an ActionListener. This means they can be clicked
on. It's mostly because I don't want to have to deal with the hassle of
buttons, because they can be a lot more complicated.
- Labels also now render a background. It just makes me happier this
way.
- Sprite now has a way to load sprites directly from a path, skipping
the sprite sheet (well, only practically skipping the sheet. They are
technically still using one). I chose this approach over actual sprite
sheets because I wanted to. And because the way Im drawing sprites
requires them to be more versatile, i.e potentially different sizes,
since I am supporting multiple sprite textures and implemented this
variance. This is no tilebased system.
- Components also now capture Mouse Scroll events. Guess why
- Fixed a minor bug with textfields- if you had a text field on a panel
that was in turn in another panel, the nested panel wouldnt get the
necessary key type events since it isnt focusable. So I made an
exception for that. My code is really exceptional! in a bad way because
Im making lots of exceptions to things.

But, atleast Im still learning. In my post mortem for my Platty project
Im going to include everything I've learned when it comes to
implementing these libraries that Im patching over time, and
spaghettiness.