Minor changes/additions
- UIPanels now have a method to clear the focus from all key-event
receiving components.
- UIPanels now have a method to get their color. This should go under
UIComponent, but oh well.
- I expanded the range of the keys boolean array to the max Short value
plus 1. I got tired of exceptions being thrown when I pressed keys out
of range so now it is a looooot bigger than necessary.
- Added a variation on the parseInt method from Integer, to the
MathUtils class. The whole idea is that entering illegal strings will
return 0, to avoid exceptions.
- Added a stringUtils class, which taught me that you can't extend
String. After a brief, sad moment I built a method to set all characters
of a string to lower case, but the first one to upper case.
- UITextfield now has a setFont method, to change font. It also offers
in-line support to creating new UITextFields. It is beyond me why I
didnt include that in the constructor in the first place, when I wrote
this for another project, but I can only assume I just didn't care about
versatility.
- UITextField also now adjusts the cursor offset appropriately when
focus is given. This way, you don't have to type anything first to
update it. Really just a visual thing.